Add upvote/downvote to change quality
Added JS, routes/controller, and tests to change quality of ideas.

The code is currently a little messy and requires changing values in 3 places,
and upvote/downvote are handled separately (at least initially).

As well, enums were finicky and required adding a new method to the controller to
handle strings being added to the db since it wasn't being converted to integers.
Strangely the request specs were able to handle that portion of it, but despite
making sure all calls were being made in JSON format, it did not perform that way
in development.